+++ app/server/rest/model/response/account.go
@@ -3,7 +3,7 @@ package response
 import "github.com/jsperandio/transaction/app/domain/model"
 
 type Account struct {
-	ID             int    `json:"account_id"`
+	ID             int64  `json:"account_id"`
 	DocumentNumber string `json:"document_number"`
 }
